requirements.txt
fastapi==0.100.1
uvicorn==0.22.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
# Quantum/TF stack pinned for KerasLayer compatibility
pennylane==0.34.0
autoray==0.6.10
tensorflow==2.13.1
keras==2.13.1
typing-extensions==4.5.0
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
redis==5.0.1
celery==5.3.4
pydantic==1.10.13
jinja2==3.1.2
aiofiles==23.2.1
